appServicePython/requirements.txt
Flask
requests
datetime
python-dateutil